requirements.txt
# Core PepperAI Dependencies
flask>=2.3.0
flask-sqlalchemy>=3.0.0
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# Quality Analysis
scikit-image>=0.20.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Disease Detection (Optional - for advanced features)
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
albumentations>=1.3.0
